fixes folder structure, fixes handlebars css, cleans up project directory, fixes public resources
