--- app/config/connection.js
@@ -1,15 +1,0 @@
-var Sequelize = require("sequelize");
-
-
- var sequelize = new Sequelize ("friendify_db", "root", "root", {
-    host: "localhost", 
-    port: 3306, 
-    dialect: "mysql", 
-    pool: {
-        max: 5, 
-        min: 0, 
-        idle: 10000
-    }
-});
-
-module.exports = sequelize;
