+++ models/friend.js
@@ -11,9 +11,10 @@
 var Sequelize = require("sequelize");
 // sequelize (lowercase) references my connection to the DB.
 var sequelize = require("../config/connection.js");
+// module.exports = function(application, req, res){
 
-var Friend = sequelize.define("Friend", {
-    // profilePhoto: DataTypes.BLOB,
+var Friend = sequelize.define("friends", {
+    profilePhoto: Sequelize.BLOB('long'),
     firstName: Sequelize.STRING,
     lastName: Sequelize.STRING,
     email: Sequelize.STRING,
@@ -25,8 +26,6 @@ var Friend = sequelize.define("Friend", {
     summary: Sequelize.TEXT,
     interests: Sequelize.TEXT
 });
-
-Friend.sync();
+Friend.sync({force: false});
 
 module.exports = Friend;
-
