+++ public/author-manager.html
@@ -13,7 +13,7 @@
      <!-- Icon Bar (Sidebar - hidden on small screens) -->
      <nav class="w3-sidebar w3-bar-block w3-small w3-hide-small w3-center">
        <!-- Avatar image in top left corner -->
-       <img id="navProfileImg" src="../assets/images/earth.jpg" class="profile-image img-circle">
+       <img id="navProfileImg" src="images/earth.jpg" class="profile-image img-circle">
        <hr>
      
        <div class="btn-group dropdown">
