+++ public/signIn.html
@@ -52,13 +52,13 @@
                 
                 <div class="form-group">
                 <label for="lblUserEmail">Email: </label>
-                <input type="email" class="form-control" id="exampleFormControlInput1" placeholder="name@example.com">
+                <input type="email" class="form-control" id="email" placeholder="name@example.com">
                 </div>
 
                 <div class="form-group">
                 <label for="lblUserPw">Password: </label>
                 
-                <input type="password" name ="password" class="form-control txtUserPw" id="exampleFormControlInput1" placeholder= "password">
+                <input type="password" name ="password" class="form-control txtUserPw" id="password" placeholder= "password">
                 </div>
                 
                 <button type="submit" class="btn btn-primary">Sign In</button>
