+++ views/layouts/main.handlebars
@@ -4,7 +4,7 @@
 <head>
   <meta charset="utf-8">
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-beta.3/css/bootstrap.min.css">
-  <link rel="stylesheet" href="../../public/css/styles.css" media="screen" title="no title" charset="utf-8">
+  <link rel="stylesheet" href="/css/styles.css" media="screen" title="no title" charset="utf-8">
   <title>Home Page</title>
 </head>
